requirements.txt
alembic
celery
compress_pickle
coverage
Cython
flake8
flake8-docstrings
flake8-import-order
flask-restx
Flask
Flask-Caching
Flask-SQLAlchemy
git+https://github.com/okankop/vidaug
jinja-compose
kaldi-io
line_profiler
locust
matplotlib
networkx
numpy
redis
requests
pandas
psycopg2
pytest
pytest-mock
seaborn
sklearn
sqlalchemy-utils==0.36.1
tqdm
watchdog==0.9
werkzeug==0.16.1
